Crate starry_api

Re-exports

pub use backend::*;

pub use utils::*;

Modules

backend

utils

Structs

Tms

utimbuf

Functions

clear_proc_sem

Called when a process exits to perform undo operations

clear_proc_shm

do_exit

exit_robust_list

sys_accept

sys_access

sys_arch_prctl

sys_bind

sys_brk

sys_chdir

sys_chmod

sys_chown

sys_clock_gettime

sys_clone

sys_close

sys_connect

sys_dup

sys_dup2

sys_epoll_create1

sys_epoll_ctl

sys_epoll_wait

sys_execve

sys_exit

sys_exit_group

sys_faccessat2

sys_fchmod

sys_fchmodat

sys_fchown

sys_fchownat

sys_fcntl

sys_fdatasync

sys_fork

sys_fstat

Get file metadata by fd and write into statbuf.

sys_fstatat

sys_fstatfs

sys_fsync

sys_ftruncate

sys_futex

sys_get_robust_list

sys_getcwd

sys_getdents64

sys_getegid

sys_geteuid

sys_getgid

sys_getitimer

Get interval timer value

sys_getpeername

sys_getpid

sys_getppid

sys_getrandom

sys_getrlimit

sys_getrusage

sys_getsockname

sys_gettid

sys_gettimeofday

sys_getuid

sys_ioctl

The ioctl() system call manipulates the underlying device parameters of special files.

sys_kill

sys_lchown

sys_link

sys_linkat

create a link from new_path to old_path old_path: old file path new_path: new file path flags: link flags return value: return 0 when success, else return -1.

sys_listen

sys_lseek

sys_lstat

Get the metadata of the symbolic link and write into buf.

sys_mkdir

sys_mkdirat

sys_mmap

sys_mount

sys_mprotect

sys_msgctl

sys_msgget

sys_msgrcv

sys_msgsnd

sys_msync

sys_munmap

sys_nanosleep

Sleep some nanoseconds

sys_open

Open a file by filename and insert it into the file descriptor table.

sys_openat

Open or create a file. fd: file descriptor filename: file path to be opened or created flags: open flags mode: see man 7 inode return new file descriptor if succeed, or return -1.

sys_pipe2

sys_poll

sys_ppoll

sys_pread64

sys_preadv

sys_preadv2

sys_prlimit64

sys_pselect6

sys_pwrite64

sys_pwritev

sys_pwritev2

sys_read

Read data from the file indicated by fd.

sys_readlink

sys_readlinkat

sys_readv

sys_recvfrom

sys_rename

sys_renameat

sys_renameat2

sys_rmdir

sys_rt_sigaction

sys_rt_sigpending

sys_rt_sigprocmask

sys_rt_sigqueueinfo

sys_rt_sigreturn

sys_rt_sigsuspend

sys_rt_sigtimedwait

sys_rt_tgsigqueueinfo

sys_sched_getaffinity

sys_sched_getscheduler

sys_sched_getscheduler_max

sys_sched_getscheduler_min

sys_sched_setaffinity

sys_sched_setscheduler

sys_sched_yield

sys_select

sys_semctl

Control operations on semaphores

sys_semget

Get a semaphore set identifier

sys_semop

Perform operations on semaphores

sys_sendfile

sys_sendto

sys_set_robust_list

sys_set_tid_address

To set the clear_child_tid field in the task extended data.

sys_setitimer

Set interval timer value

sys_setrlimit

sys_shmat

sys_shmctl

sys_shmdt

sys_shmget

sys_sigaltstack

sys_socket

sys_socketpair

sys_stat

Get the file metadata by path and write into statbuf.

sys_statfs

sys_statx

sys_symlink

sys_symlinkat

sys_sysinfo

sys_syslog

sys_tgkill

sys_times

sys_tkill

sys_umount2

sys_uname

sys_unlink

sys_unlinkat

remove link of specific file (can be used to delete file) dir_fd: the directory of link to be removed path: the name of link to be removed flags: can be 0 or AT_REMOVEDIR return 0 when success, else return -1

sys_utime

sys_utimensat

sys_utimes

sys_wait4

sys_write

Write data to the file indicated by fd.

sys_writev