starry_api

Function sys_ioctl

pub fn sys_ioctl(
    _fd: i32,
    _op: usize,
    _argp: UserPtr<c_void>,
) -> LinuxResult<isize>

The ioctl() system call manipulates the underlying device parameters of special files.

Arguments

fd - The file descriptor

op - The request code. It is of type unsigned long in glibc and BSD, and of type int in musl and other UNIX systems.

argp - The argument to the request. It is a pointer to a memory location