starry_api::socket

Trait SocketAddrExt

pub trait SocketAddrExt: Sized {
    // Required methods
    fn read_from_user(
        addr: UserConstPtr<sockaddr>,
        addrlen: socklen_t,
    ) -> LinuxResult<Self>;
    fn write_to_user(&self, addr: UserPtr<sockaddr>) -> LinuxResult<socklen_t>;
    fn family(&self) -> u16;
    fn addr_len(&self) -> socklen_t;
}

Trait to extend SocketAddr and its variants with methods for reading from and writing to user space.

Required Methods

fn read_from_user( addr: UserConstPtr<sockaddr>, addrlen: socklen_t, ) -> LinuxResult<Self>

This method attempts to interpret the data pointed to by addr with the given addrlen as a valid socket address of the implementing type.

fn write_to_user(&self, addr: UserPtr<sockaddr>) -> LinuxResult<socklen_t>

This method serializes the current socket address instance into the [sockaddr] structure pointed to by addr in user space.

fn family(&self) -> u16

Gets the address family of the socket address.

fn addr_len(&self) -> socklen_t

Gets the encoded length of the socket address.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.