Crate starry_api

Modules

file

path

ptr

signal

sockaddr

Wrapper for sockaddr. Adapted from rustix::net::SocketAddrAny.

time

Structs

Tms

Enums

FileType

Functions

check_mounted

check if a path is mounted

do_exit

mount_fat_fs

Mount a fatfs device

sys_arch_prctl

sys_brk

sys_chdir

sys_clock_gettime

sys_clone

sys_close

sys_dup

sys_dup2

sys_execve

sys_exit

sys_exit_group

sys_fcntl

sys_fork

sys_fstat

Get file metadata by fd and write into statbuf.

sys_fstatat

sys_futex

sys_getcwd

sys_getdents64

sys_getegid

sys_geteuid

sys_getgid

sys_getpid

sys_getppid

sys_gettid

sys_gettimeofday

sys_getuid

sys_ioctl

The ioctl() system call manipulates the underlying device parameters of special files.

sys_kill

sys_link

sys_linkat

create a link from new_path to old_path old_path: old file path new_path: new file path flags: link flags return value: return 0 when success, else return -1.

sys_lseek

sys_lstat

Get the metadata of the symbolic link and write into buf.

sys_mkdirat

sys_mmap

sys_mount

sys_mprotect

sys_munmap

sys_nanosleep

Sleep some nanoseconds

sys_open

Open a file by filename and insert it into the file descriptor table.

sys_openat

Open or create a file. fd: file descriptor filename: file path to be opened or created flags: open flags mode: see man 7 inode return new file descriptor if succeed, or return -1.

sys_pipe2

sys_read

Read data from the file indicated by fd.

sys_readv

sys_rt_sigaction

sys_rt_sigpending

sys_rt_sigprocmask

sys_rt_sigqueueinfo

sys_rt_sigreturn

sys_rt_sigsuspend

sys_rt_sigtimedwait

sys_rt_tgsigqueueinfo

sys_sched_yield

sys_set_tid_address

To set the clear_child_tid field in the task extended data.

sys_sigaltstack

sys_stat

Get the file metadata by path and write into statbuf.

sys_statx

sys_tgkill

sys_times

sys_tkill

sys_umount2

sys_uname

sys_unlink

sys_unlinkat

remove link of specific file (can be used to delete file) dir_fd: the directory of link to be removed path: the name of link to be removed flags: can be 0 or AT_REMOVEDIR return 0 when success, else return -1

sys_waitpid

sys_write

Write data to the file indicated by fd.

sys_writev

umount_fat_fs

unmount a fatfs device